requirements.txt
streamlit>=1.30.0,<2.0
python-dotenv>=1.0.0,<2.0
keybert>=0.6.0,<1.0
openai>=1.0.0,<2.0
jieba>=0.42,<1.0
selenium>=4.20.0,<5.0
requests>=2.30,<3.0
beautifulsoup4>=4.12,<5.0
scikit-learn>=1.3,<2.0
